initial tensorflow support, untested, not model yet
