--- bindle/models/invoice.toml
@@ -1,26 +1,0 @@
-bindleVersion = "1.0.0"
-
-[bindle]
-name = "identity_model"
-version = "0.1.0"
-authors = ["Christoph Brewing <finnfalter@gmail.com>"]
-description = "My first bindle"
-
-[annotations]
-myname = "myvalue"
-
-
-[[parcel]]
-[parcel.label]
-sha256 = 'a21f9949bc87bb2eaf8508872be52ed1b6aaf4b4f8d546ac8254777818c53b09'
-mediaType = 'application/json'
-name = 'identity_input_output.json'
-size = 189
-
-
-[[parcel]]
-[parcel.label]
-sha256 = 'eb68abfa54af414a6aa8ce2f7554c0076ad3833149b4fb5d99538578213fffd9'
-mediaType = 'application/octet-stream'
-name = 'identity_input_output.onnx'
-size = 112
