+++ bindle/scripts/bindle_create.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+
+source $(dirname ${BASH_SOURCE[0]})/../../deploy/env
+
+# in case there are multiple models to push append lines like the next one
+source ./push_bindle.sh ../models/identity_input_output.toml ../models/identity_input_output.csv
