+++ providers/mlinference/Cargo.toml
@@ -17,9 +17,10 @@ serde_json = "1.0"
 thiserror  = "1.0"
 tokio = { version = "1", features = ["full"] }
 toml = "0.5"
-tract-data       = "0.16.1"
-tract-linalg      = "0.16.1"
-tract-onnx  = { version = "0.16.1" }
+tract-data = "0.16.1"
+tract-linalg = "0.16.1"
+tract-onnx = { version = "0.16.1" }
+tract-tensorflow = "0.16.1"
 wasmbus-rpc = "0.8.3"
 wasmcloud_interface_mlinference = { path="../../interfaces/mlinference/rust" }
